Uses the sync API v1 fro syncing.
For the syncing, the app no longer calls the weird-homemade-frankenstein-looking xml tha called before, but the sync API v1. As of now, it updates the versions of the different database sections, but it doesn't use them as reference (still uses the old preference value). Also, the actual sync data is received, but it still needs to be parsed and inserted into the database.