DB tables checked every time.
Before, the existence of the database tables was checked only on the first run. Adding new tables to the model implies that now they must be checked every time. If they exists, nothing will happen. Also, some improvements on the API v1 sync process, but db content syncing is still not ready.