Improvements in sections and entries.
Styles have been defined for sections and entries, so they look more like they do en the website and more uniform and clean. For the home section, all of them have been styled, and their margins and padding have been set. For some dinamically loaded entries, margins are applied programmatically, and the values have been established in the GM class. Some of those changes have produced diferent results in other sections of the app that will have to be reviewed.