Implemented changelog.
Changelog entry is now working. Clicking the entry from the settings screen shows a dialog with a changelog, version by version. The changelog is an html file inside the app. Three of them are present, one for each language, and the appropiate one is loaded. However, as of now, all three are identical and in spanish.